+++ NEWS.md
@@ -4,6 +4,7 @@
 * Removed `as_tibble.grouped_df()` to prevent masking `dplyr::as_tibble.grouped_df()`.
 * Added `summary()` method for time classes, allowing `summary()` to be called on tsibble objects. (#319)
 * `select()` and `transmute()` now warn when the index variable is dropped and automatically re-added.
+* `tsibble()` now builds columns sequentially, allowing columns to refer to previously created columns in the call. (#289)
 * Documentation improvements.
 
 # tsibble 1.1.6